refactor: Update all migration scripts to change integer IDs to BIGINT for improved scalability

+++ master/src/main/resources/db/migration/V10__create_rooms_amenities.sql
@@ -1,6 +1,6 @@
 CREATE TABLE rooms_amenities (
-    id_room INTEGER NOT NULL,
-    id_amenity INTEGER NOT NULL,
+    id_room BIGINT NOT NULL,
+    id_amenity BIGINT NOT NULL,
 
     PRIMARY KEY (id_room, id_amenity),
